refactor(#3777): split cqlite-cli json.rs unit tests into json_tests.rs (campsite rule)
